backend_os/requirements.txt
asgiref
certifi
cloudinary
Django
django-cors-headers
django-filter
djangorestframework
Markdown
pytz
sqlparse
urllib3